app/pyproject.toml
[tool.poetry]
name = "example air pollution rest api"
version = "0.1.0"
description = "A example REST API serving air pollution data."
authors = ["Martin Nigbur <martin.nigbur@mhp.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
fastapi-cache2 = "^0.2.2"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
httpx = "^0.27.2"
influxdb-client = "^1.46.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
